mechaduino_function stuff to new locations and deleted
