--- Mechaduino_01/Mechaduino_01/Mechaduino_functions.cpp
@@ -1,59 +1,0 @@
-// All Mechaduino functions are in this file
-
-#include "Arduino.h"
-#include <Wire.h>
-
-
-
-
-//interrupt vars
-
-extern volatile float ei;
-extern volatile int U;  //control effort (abs)
-extern volatile float r;  //setpoint
-extern volatile float y;  // measured angle
-extern volatile float yw;
-extern volatile float yw_1;
-extern volatile float e;  // e = r-y (error)
-extern volatile float p;  // proportional effort
-extern volatile float i;  // integral effort
-extern volatile float PA;  //
-
-extern volatile float u;  //real control effort (not abs)
-extern volatile float u_1;
-extern volatile float e_1;
-extern volatile float u_2;
-extern volatile float e_2;
-extern volatile float u_3;
-extern volatile float e_3;
-extern volatile long counter;
-
-extern volatile long wrap_count;
-extern volatile float y_1;
-
-
-
-
-extern volatile float ITerm;
-
-extern volatile char mode;
-
-
-
-
-
-
-
-
-
-void receiveEvent(int howMany)
-{
-  while (1 < Wire.available()) // loop through all but the last
-  {
-    char c = Wire.read(); // receive byte as a character
-    SerialUSB.print(c);         // print the character
-  }
-  int x = Wire.read();    // receive byte as an integer
-  SerialUSB.println(x);         // print the integer
-  r = 0.1 * ((float)x);
-}
