--- Mechaduino_01/Mechaduino_01/Mechaduino_functions.h
@@ -1,13 +1,0 @@
-//Mechaduino functions header file
-
-
-#ifndef Mechaduino_functions
-#define Mechaduino_functions
-
-        
-    void receiveEvent(int howMany);
-	
-	
-	
-#endif
-
