+++ Mechaduino_01/Mechaduino_01/State.cpp
@@ -4,6 +4,4 @@
 #include "Parameters.h"
 #include "Utils.h"
 
-extern int dir;
-extern int stepNumber;
-extern void output(float,int);
+
