+++ Mechaduino_01/Mechaduino_01/State.h
@@ -3,6 +3,40 @@
 #ifndef __STATE_H__
 #define __STATE_H__
 
+//interrupt vars
+
+extern volatile float ei;
+extern volatile int U;  //control effort (abs)
+extern volatile float r;  //setpoint
+extern volatile float y;  // measured angle
+extern volatile float yw;
+extern volatile float yw_1;
+extern volatile float e;  // e = r-y (error)
+extern volatile float p;  // proportional effort
+extern volatile float i;  // integral effort
+extern volatile float PA;  //
+
+extern volatile float u;  //real control effort (not abs)
+extern volatile float u_1;
+extern volatile float e_1;
+extern volatile float u_2;
+extern volatile float e_2;
+extern volatile float u_3;
+extern volatile float e_3;
+extern volatile long counter;
+
+extern volatile long wrap_count;
+extern volatile float y_1;
+
+
+extern volatile float ITerm;
+
+extern volatile char mode;
+
+
+
+
+
 extern int dir;
 extern int stepNumber;
 extern void output(float,int);
